Implemented new approach for fuzzer synchronization
This commit changes the way fuzzer workers synchronize. Previously, a new
worker would be sent the entire set of interesting samples every discovered by
the responsible master. It would then execute all of them and thus be
synchronized. This had the advantage of filtering out non-deterministic samples
but obviously was very slow for large corpuses.
With this commit, the fuzzer master now exports its internal state consisting of
the state of its evaluator (e.g. the coverage bitmap for edge coverage) and the
current (active) corpus. It then sends both to the worker which simply imports
them without having to execute any programs. The worker is then basically a
clone of the master. Afterwards, workers will (hopefully) diverge from the
master due to dropout and non-deterministic samples.
This now also makes it possible to limit the size of the corpus, e.g. to avoid
OOMing if too many intersting samples are discovered.

+++ Sources/libsocket/include/libsocket.h
@@ -22,8 +22,8 @@ int socket_listen(const char* address, uint16_t port);
 int socket_accept(int socket);
 int socket_connect(const char* address, uint16_t port);
 
-int socket_send(int socket, const uint8_t* data, uint32_t length);
-int socket_recv(int socket, uint8_t* buffer, uint32_t length);
+long socket_send(int socket, const uint8_t* data, long length);
+long socket_recv(int socket, uint8_t* buffer, long length);
 
 int socket_close(int socket);
